Bon voila le buffer est un peu implemente, mais il reste encore pas mal de choses a faire

+++ libft/ft_nbrlen.c
@@ -6,13 +6,13 @@
 /*   By: snunes <marvin@42.fr>                      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/05/29 15:34:52 by snunes            #+#    #+#             */
-/*   Updated: 2019/05/29 15:41:07 by snunes           ###   ########.fr       */
+/*   Updated: 2019/06/07 13:12:06 by snunes           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_nbrlen(int nbr)
+int	ft_nbrlen(long long int nbr)
 {
 	int len;
