+++ libft/include/libft.h
@@ -6,7 +6,7 @@
 /*   By: snunes <snunes@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/04/08 19:07:48 by snunes            #+#    #+#             */
-/*   Updated: 2019/06/05 05:18:58 by snunes           ###   ########.fr       */
+/*   Updated: 2019/06/07 13:12:29 by snunes           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -37,8 +37,8 @@ typedef	struct	s_list
 }				t_list;
 
 void			ft_putlongnbr(long long int nbr);
-int				ft_nbrrev(int nbr);
-int				ft_nbrlen(int nbr);
+long long int	ft_nbrrev(long long int nbr);
+int				ft_nbrlen(long long int nbr);
 void			ft_strrev(char *to_reverse);
 int				ft_sqrt(int nb);
 int				ft_lst_length(t_list *lst);
